Metadata Hedging: Refactors helper types into MetadataHedgingStrategy and locks 1.5s threshold
- Fold MetadataHedgeSkipReason, HedgeBranch, MetadataHedgeEligibility,
  MetadataHedgeDiagnostics, MetadataHedgingResult and MetadataHedgingContext
  into MetadataHedgingStrategy as nested types; delete the standalone files.
- Fold RetryUtility.IsRegionalFailure into MetadataHedgingStrategy as an
  internal static helper; delete RetryUtility.cs and rename its tests.
- Remove the internal MetadataHedgingOptions class; the hedge threshold is now
  a fixed SDK-derived value (FirstAttemptTimeout + 500ms = 1.5s) and the
  concurrency budget is a constant, so the 1.5s cold-start threshold is always
  honored and not customer-tunable.
- Update production and test references; all metadata-hedging unit tests pass.
