--- Microsoft.Azure.Cosmos/src/MetadataHedgingOptions.cs
@@ -1,46 +1,0 @@
-//------------------------------------------------------------
-// Copyright (c) Microsoft Corporation.  All rights reserved.
-//------------------------------------------------------------
-
-namespace Microsoft.Azure.Cosmos
-{
-    using System;
-
-    /// <summary>
-    /// Tuning knobs for cold-start metadata cache hedging. Used by
-    /// <see cref="Cosmos.Routing.MetadataHedgingStrategy"/> when
-    /// <see cref="CosmosClientOptions.EnableMetadataHedgingForColdStart"/> is
-    /// effectively on.
-    /// </summary>
-    /// <remarks>
-    /// See <c>docs/PPAF_Metadata_Hedging_ColdStart_Design.md</c> §5.1.
-    /// </remarks>
-    internal sealed class MetadataHedgingOptions
-    {
-        /// <summary>
-        /// Default per-client concurrency budget for in-flight metadata hedges.
-        /// Mirrors the design §5.11 default.
-        /// </summary>
-        public const int DefaultPerClientConcurrencyBudget = 8;
-
-        /// <summary>
-        /// Added to <c>HttpTimeoutPolicy.FirstAttemptTimeout</c> to derive the
-        /// default hedge <see cref="Threshold"/> when the caller does not supply
-        /// one. Default: 500&#160;ms.
-        /// </summary>
-        public static readonly TimeSpan DefaultThresholdStep = TimeSpan.FromMilliseconds(500);
-
-        /// <summary>
-        /// Time after which the hedge branch is dispatched if the primary has
-        /// not produced an acceptable response. When <c>null</c>, defaults to
-        /// <c>HttpTimeoutPolicy.FirstAttemptTimeout + 500&#160;ms</c> (today
-        /// 1.5&#160;s).
-        /// </summary>
-        public TimeSpan? Threshold { get; set; }
-
-        /// <summary>
-        /// Per-client cap on in-flight metadata hedges. Default: 8.
-        /// </summary>
-        public int PerClientConcurrencyBudget { get; set; } = DefaultPerClientConcurrencyBudget;
-    }
-}
